# swagger.yaml
openapi: 3.0.0
info:
  title: TODOEVENTS
  version: 1.0.0
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    noauthAuth:
      type: http
      scheme: noauth
paths:
...
  /api/v1/events:
    get:
      tags:
        - default
      summary: get_event_by_name
      description: >
        Generated from cURL: curl -X GET
        "http://localhost:5000/api/v1/events?search=Tech"
      parameters:
        - name: search
          in: query
          schema:
            type: string
          example: Reactive
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    post:
      tags:
        - default
      summary: create_events
      description: |
        Generated from cURL: curl -X POST http://localhost:5000/api/v1/events \
          -H "Content-Type: application/json" \
          -H "Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJmcmVzaCI6ZmFsc2UsImlhdCI6MTc1NDE2NDg1NCwianRpIjoiNjNiYzZmNDktNzhkNi00MjYxLTkyOWYtNzU4NzkyZWE2MDIwIiwidHlwZSI6ImFjY2VzcyIsInN1YiI6NSwibmJmIjoxNzU0MTY0ODU0LCJjc3JmIjoiNTYxZTAwNjMtODQ2Ny00OWVmLTk1NjItMDk0ZmRlNjkyM2Q0IiwiZXhwIjoxNzU0MTY4NDU0LCJlbWFpbCI6Imp1YW5pdGFAdGVzdC5jb20iLCJyb2xlIjoib3JnYW5pemVyIn0.OD-AK34Rjvr6AaF1XCd7opvYCJ-aYXyOvzIiMfAxz30" \
          -d '{
            "name": "AI & Cloud Conference 2025",
            "description": "Annual event about AI and cloud technologies",
            "capacity": 200,
            "status": "open"
          }'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                name: reactive programming Conference 2025
                description: Annual event reactive programming
                capacity: 200
                status: published
                start_at: '2025-08-15T09:00:00'
                end_at: '2025-08-15T18:00:00'
      security:
        - noauthAuth: []
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: Authorization
          in: header
          schema:
            type: string
          example: >-
            Bearer
            eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJmcmVzaCI6ZmFsc2UsImlhdCI6MTc1NDIzNzM0MiwianRpIjoiYjVhZmI1N2YtOGI0Zi00ZTMzLTk3MTItNWIyZTE3Mzk0NjY3IiwidHlwZSI6ImFjY2VzcyIsInN1YiI6IjEiLCJuYmYiOjE3NTQyMzczNDIsImNzcmYiOiIxNzQ1NTYzYS03MmUzLTRhMDQtODAyMS0zMDJlZWQwMWU3YWMiLCJleHAiOjE3NTQzMjM3NDIsImVtYWlsIjoianVsaWFuQHRlc3QuY29tIiwicm9sZSI6Im9yZ2FuaXplciJ9.XWghJ7UvedM6qnm2190rlIlwjbrhEJQTONanPfwz-_M
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...